♻️ component하위의 페이지 폴더를 src하위로 뽑고 페이지별로 폴더를 만듦
