--- src/components/pages/RankingPage.jsx
@@ -1,15 +1,0 @@
-import CalendarBar from "../ui/Ranking/CalendarBar";
-import RecordArea from "../ui/Ranking/RecordArea";
-import TopThreeArea from "../ui/Ranking/TopThreeArea";
-
-const RankingPage = () => {
-  return (
-    <>
-      <CalendarBar />
-      <TopThreeArea />
-      <RecordArea />
-    </>
-  );
-};
-
-export default RankingPage;
