+++ src/pages/Ranking/RankingPage.jsx
@@ -1,0 +1,15 @@
+import CalendarBar from "@/pages/Ranking/CalendarBar.jsx";
+import RecordArea from "@/pages/Ranking/RecordArea.jsx";
+import TopThreeArea from "@/pages/Ranking/TopThreeArea.jsx";
+
+const RankingPage = () => {
+  return (
+    <>
+      <CalendarBar />
+      <TopThreeArea />
+      <RecordArea />
+    </>
+  );
+};
+
+export default RankingPage;
