+++ src/pages/Ranking/RecordArea.jsx
