+++ src/pages/SelectHobbies/SelectHobbiesPage.jsx
